💄 ♻️ moved identity menu into its own component, status badge and identity menu switched position
