+++ src/app/app.component.ts
@@ -21,30 +21,6 @@ import { take } from 'rxjs';
 export class AppComponent implements OnInit {
   menuItems: NavMenuItem[] = [];
 
-  profileItems: NavMenuItem[] = [
-    {
-      label: $localize`User Settings`,
-      icon: faUserGear,
-      routerLink: ['user'],
-    },
-    {
-      label: $localize`Edit Account`,
-      command: () => {
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        const iss = this.oauthService.getIdentityClaims().iss;
-        const referrer = encodeURI(this.appConfigService.getConfiguration().clientId);
-        const referrerUri = `${encodeURI(window.location.href)}`;
-        window.location.href = `${iss}/account?referrer=${referrer}&referrer_uri=${referrerUri}`;
-      },
-      icon: faUser,
-    },
-    {
-      label: $localize`Sign out`,
-      icon: faSignOut,
-      command: () => this.oauthService.logOut(),
-    },
-  ];
   name = '';
   username = '';
   avatarSrc = '';
